Set action on recyclerview in Porduces
it opens a new activity for viewing/editing/deleting the item
fragment for adding new item

+++ Client/app/src/main/java/com/example/client/data/model/Product.java
@@ -2,5 +2,5 @@ package com.example.client.data.model;
 
 import java.util.Date;
 
-public record Product(int id, String name, Date expirationDate, int quantity, String container) {
+public record Product(long id, String name, Date expirationDate, float quantity, String container) {
 }
